feat(poweredBy): getWidgetRenderState (#4551)
* feat(poweredBy): getWidgetRenderState

DX-206

* fix jsdoc
